Gradebook
A program that displays student grade information. It will also allow you to update the information for an existing student, or add new student. Additionally, it should have a new view that displays a picture of the student along with that student's name and address and one that allows you to add or update a student's info.